Updated openstack/openstack Project: openstack/nova  fc1076831991aaa2c07556ea855ccea3e73f6aa7
More specific error messages on building BDM

When a user failed to create an instance with BDM because of
exceeded quota, its error message should be one that contains
the reason rather than generic 'Failure prepping block device'.

For this purpose, we catch VolumeLimitExceeded explicitly in
_build_resources(). As a result, the user will get a clearer
message 'Volume resource quota exceeded'.

We catch InvalidBDM explicitly as well. With this, although its
error message ('Block Device Mapping is Invalid.') is still not
user-friendly, it gets rid of a meaningless stack trace in log;
operator-friendly.

Change-Id: Ia72a3cdde640f8e18fe3778e903204930e96d402
Closes-bug: #1482191