Updated openstack/openstack Project: openstack/tempest  12b47cbe798bd8df40e8404da3ea6873b30de022
Handle 'error_restoring' status in wait_for_volume_status

If cinder hits an error when restoring a volume from a backup the volume
is set to 'error_restoring' status. If tempest doesn't handle this and
fail fast, tests keep waiting for the volume to go to 'available' status
or timeout - and the timeout is what happens.

Related-Bug: #1483434

Change-Id: I8f6d0828a84ec5d4a5fa6332e55152bd67eaa605