Updated openstack/openstack Project: openstack/ceilometer  6a7d3feb476bd1e6fe2ab78fbe87245e31f03985
start rpc deprecation

we switch to work queues in Juno and made it the default rather
than rpc. the main advantage of using RPC is it offers callback
functionality at the expense of performance. we don't need callback
for our use cases so it's best to stop supporting rpc and leave just
udp and queues in our collector.

Change-Id: Ia02421ae09cdf27998ce76a99170100f36595f71
Related-Bug: #1479820