Updated openstack/openstack Project: openstack/nova  1b92f480a970c3944c6de33ef838936d2badd670
Use string substitution before raising exception

Fixes a number of places where an exception was raised where it is
passed both a message argument and kwargs.  Since the exceptions are
ultimately NovaException objects, the kwargs are never substituted into
the message because message is used instead of msg_fmt.  Simply
substituting the kwargs into the message before raising the exception
fixes the issue.

Closes-Bug: #1483022
Change-Id: I62180fcbb0ee3385ec3d271f30de58eebaf62237