Updated openstack/openstack Project: openstack/nova  017c9523f735237963e60b4a82f7d00deb7d8497
libvirt: qemu-img convert should be skipped when migrating

Image conversion is done to migrate the complete qcow2 image.
But, it is unnecessary to merge the backing file.
I have modified so that nova-compute just copy
the file instead of conversion.
It is possible to move the file in case of
a shared storage environment, but nova resize-revert
fails in that case.

Closes-Bug: #1399098
Change-Id: Ic05b48761f09a4711a11bacfc64cbcf39b352e84