Updated openstack/openstack Project: openstack-dev/devstack  e8c70e23b5c8b9a41e2e86116972cf4da3367e19
Tempest: add a Cinder extend_with_snapshot feature flag

A new tempest test is being added in https://review.openstack.org/#/c/200108/
but it doesn't run by default because the test fails on Juno. So a
feature flag in Tempest is added. This patch turns on this feature
flag at Devstack's side.

Change-Id: If1cf90dac3edc81a483fc51da74495042c96d543