Updated openstack/openstack Project: openstack/governance  da41f1fe95578144c0a3617a4d5c7e9e115e8b5d
Adds Debian & derivatives packaging to OpenStack

We would like to have the packaging Git repositories for Debian
hosted in the /openstack namespace.

As per discussions, the RPM folks would be packaging separately, and
we wouldn't form a single team.

Within the projects, the following vendors would be involved:
- Debian contributors
- MOS Linux (for the Debian part of it)
- Maybe H-Linux if they also want to participate

This commit contains an initial list of package repositories. It
is made currently only of server packages. Packages for Oslo
libraries and clients are currently out of scope of this commit,
however they may appear in the near future.

We already have plans to:
- build automatically, with parts of the tooling already designed.
- gate on each commit using standard distribution tools like
piuparts, lintian, adequate and so on.

The core team is partially defined, and composed with at least:
- Thomas Goirand (PTL for this Liberty cycle)
- Igor Yozhikov
- Ivan Udovichenko
- Daniil Trishkin

We may later on include more core reviewer from Mirantis, and this
list isn't exhaustive yet.

(Thomas asked about update of this patch-set)

Change-Id: I73b64c4bc0497a1d42b6308a038daabab258ede8