Updated openstack/openstack Project: openstack/nova  89cf4d551526a0673af00a415730df8edb80f439
Allow resize root_gb to 0 for volume-backed instances

Currently, after an instance launched with bootable
volume and flavor with non-zero disk. If we resize
this instance to an flavor with zero disk, Nova API
will throw exception.

As for volume-backed instance, the root volume doesn't
occupy any space in nova-compute node. So it should not
be limited by flavor's root_gb.

Change-Id: I1ab1a2625650a828b94c0af2147b0aa53a652a82
Closes-Bug: 1480040