Updated openstack/openstack Project: openstack/neutron-vpnaas  cb42f1d43fbe3415ba9ddd3acb7798644c886520
StrongSwanProcess: redefine DIALECT_MAP at class level

DIALECT_MAP is a class variable of BaseSwanProcess which is a base
class for StrongSwanProcess and OpenSwanProcess. Overwriting class
variables at instance init is not safe and it leads to sideefects,
i.e. when running UTs for both StrongSwanProcess and OpenSwanProcess.

DIALECT_MAP should be instead be redefined in StrongSwanProcess class
namespace.

Change-Id: I705ea4863a14fc6ecf3e0f16142e4f375c75cd62
Closes-Bug: #1477570