Updated openstack/openstack Project: openstack/ironic-specs  f281a0433076583a3f58849741fce1b6fff74f75
move supported-boot-device-list spec (1 of 2)

Due to the reorg of the specs repo, the supported-boot-device
spec needs to live in the approved/, not the liberty/,
subdirectory. (With a placeholder in the old directory.)
This move will be done in two commits to preserve the history;
git can't recognize a rename and replace operation in the same commit.

Change-Id: I424390ccae956f4ca0f9e6ba5b59ca2696cf3216