Updated openstack/openstack Project: openstack-infra/project-config  2aa62a16433e028ce3c590f762140183da869510
Don't run gate-neutron-vpnaas-dsvm-functional-sswan on stable/juno

The strongswan support was not in vpnaas in Juno so the job won't work
(devstack doesn't set it up in stable/juno).

The neutron-vpnaas repo has a stable/juno branch but it's not used so
this hasn't been a problem so far.  And neutron runs the sswan job in
the experiemental queue, so while it would fail there in stable/juno,
it's the experimental queue so it's not blocking.

This cleans things up though to avoid confusion.

Change-Id: I75c097d0d1f811e1ba8673b0003f8a5790d23a69