Updated openstack/openstack Project: openstack/ironic-specs  23435a931f6dcea9f355396040286730f055a518
Removing refrence to bind_requested flag

During Ironic mid-cycle, we decided that this
flag is not needed. Instead, we will use host-id
for this purpose. When the port does not
need to be bound, host-id will be set to None.
Neutron (ML2 plugin) already has a mechanism to
deal with this. Therefore, it was decided that
we do not need two ways to accomplish one action.

Change-Id: I7936fd7545a4672042b32e1c941798f7b5dd1272