Updated openstack/openstack Project: openstack-infra/puppet-reviewday  c006a4e12e65fac6ef0149ae53cbad94cec41fb6
Set reasonable defaults for reviewday::site

Empty string is not an appropriate default value for parameters that
are used unconditionally in the puppet manifest or template. For
example, the name of the http root directory cannot be an empty string,
so it is not sensible to use empty string as a default. This patch
attempts to set reasonable defaults for every parameter. For the key
parameters, we change the default from empty string to undef and update
the conditional logic related to those parameters.

Change-Id: I6d2e7037385a4082726e5e7db93b18a92b4a8ace