Updated openstack/openstack Project: openstack-infra/puppet-ssl_cert_check  ab05f01e16524a57fb52ff087c80f173030abd7b
Fix class parameters

All of these parameters are unconditionally fed into a cron command,
leaving them empty would result in the command failing with a syntax
error. This patch sets defaults for the email and days parameters but
makes the domainlist_file required since it is unmanaged in this
class.

Change-Id: I7a0ba47e8a1bedb08b6b9d46c36ad87cbf0243d1