Updated openstack/openstack Project: openstack-infra/puppet-nodepool  f2168cd97853d8b93c46bd968c00c55a19999452
Add creation of secure.conf file

This will be a new file that will store all the secrets needed
by nodepool at this stage: mysql password, and jenkins masters
credentials.
Following there will be a patch in nodepool, to use that file
to retrieve mysql and jenkins settings. By this way,
nodepool.yaml can be a plain file, and not a template, and can
be created into project-config.

Change-Id: Ie9381740e3644feaee1f1b201499e3a253677f39