Updated openstack/openstack Project: openstack-infra/devstack-gate  bc52a7a5e21c620908123801b2cb5ba609b583f8
Add q-vpn back to base services but exclude for liberty runs

Commit c59dfacef8c211b7e114e634ca17f225bde6c3e0 removed the q-vpn
service from being enabled in all neutron runs which is OK for liberty
since that project has it's own devstack plugin and job configs for
master, but it breaks stable/kilo.

So add the q-vpn service back in for kilo.

Related-Bug: #1473475

Change-Id: I69ea5b194fec520e1aa31eb0554cf83c447a5e18