Updated openstack/openstack Project: openstack-infra/project-config  3af2e824a06092642290b7b3d65188965fff423a
Kolla: Add a gate for image-centos-binary

This builds centos in binary mode.  The tox job uses Kolla's
next generation build script which is not compatible with the
initial version of our build script.  As a result of this we
left the existing job intact so we maintain gating of our
image building process with our initial build script tool.

As part of the new gating with the new build, we are also gating an
entirely new Dockerfile infrastructure which is templated.  These
new dockerfiles are not compatible with the legacy build script.  As
a result we need the existing kolla-functional-f21 to remain intact
until we finish the build and dockerfile templated implementations.

After that we intend to remove the kolla-function-f21 gate job,
the legacy build script and the legacy dockerfiles.

Co-Authored-By: Steven Dake <stdake@cisco.com>

Change-Id: I2a3d42d34a32a2f514b8c64f17ff54e7b08d6d7b