Updated openstack/openstack Project: openstack-infra/project-config  a115f67dea6ce2437da4ac775f19626f8da2eddb
Use zuul-cloner for bifrost jobs

bifrost does a bunch of things with multiple repos, but is currently
managing the cloning of them all internally. Get things set up to properly
use zuul-cloner to prep repo state.

This should be followed up with a bifrost patch to simply use the repos
that zuul-cloner prepares here.

Change-Id: Ied5c5287b79aadfd6660335d4c68182821e2d843