Updated openstack/openstack Project: openstack/nova  47275867a0f1fa37f20e3177b1c2f8927d91be28
Make pagination tolerate a deleted marker

Currently, Nova does not paginate correctly if the marker is
a deleted server. When trying to get servers with such a
marker, a "BadRequest" error saying that the marker is
not found will be raised.

We now always pass read_deleted=yes if a marker is being used.

Change-Id: I79a1fb79a38ac4006db622590574308aafdcb5b5
Closes-bug: #1481812
Related-Bug: #1398086