Updated openstack/openstack Project: openstack/nova  039fc0c866a9e996690df3e9bf6a652aed02395b
Store correct VirtCPUTopology

When booting a NUMA-enabled instance, the scheduler generates and
stores a meaningless VirtCPUTopology as part of the InstanceNumaCell
objects. A correct value is later generated and used in libvirt XML
generation but this is not stored. Fix this by skipping the initial
generation and storing of the VirtCPUTopology in favour of storing
the correctly generated version.

Change-Id: Ief4fcdec0e107a233225ebd68207ac6172b3751e
Closes-bug: #1466780