Updated openstack/openstack Project: openstack/keystone  343ea55a3a82b0d2a3f98d1061d7dd93ad137f66
Hardens the validated decorator's implementation

Previously the validated decorator expected to find the entities it
validated in the kwargs. Since the method signatures require the
argument I don't see any reason to force kwarg usage. This change will
now also be able to find the argument in args.

If the validated decorator was used with an incorrect resource to
validate (a misspelling, etc.), then the validated call would never
actually validate anything. This change will ensure, at import time,
that the correct resource to validate is specified.

Change-Id: Iad5f63cd34e4b75e64a1e233a0915d971a40f79a
Closes-Bug: #1473848