Updated openstack/openstack Project: openstack/keystone  7f6c3696456119d357d47272709d9bc9862d9b0d
Maintain datatypes when loading configs from DB

When loading domain configuration from database we are changing
the actual data type of the configuration to unicode string. We
want to preserve the type of the configuration value because
otherwise we have problems later in the code when for example a
'bool' configuration value like 'use_tls' is evaluated incorrectly
because its type is unicode instead of bool.

Closes-Bug: #1461299
Change-Id: I54f5bd19910528b18c428bb3702c8cd4db3cc801