Updated openstack/openstack Project: openstack/python-keystoneclient  59f815dd65317661cbd1447ae8cdd84ece81c7c6
Expose token_endpoint.Token as admin_token

When bootstrapping a cloud, using this auth plugin is necessary to be
able to provide an admin token. However, there has been some confusion
as to its officialness and the names of its parameters. Make it clear
why it's exposed, and additionally be clearer about squatting on the
name of the plugin so that things can depend on its interface.

Change-Id: I8e896b28d5879cd4b1372009926c67cde773308f