Updated openstack/openstack Project: openstack-infra/jeepyb  7f7b988b38610e677638fc92562d2c4a5c223430
Only add 'groups' file to the repo if it exists

The 'groups' file gets created only if there are groups mentioned in
the 'project.config' file.  A 'project.config' file is not required to
mention any groups, thus we should only attempt to add the 'groups' file
if it exists.

Change-Id: I333df5156bc27db29143ee412d86e12606faa751