Updated openstack/openstack Project: openstack/cinder  228b52cd0b9b0d6c6efc0a5371e7bca26c0375b4
VMware: Deprecate vCenter version less than 5.1

Log a warning if the vCenter server version is less than 5.1.
The plan is to enforce the version in M release.

DocImpact
Running Cinder with a VMware vCenter version less than 5.1 is
deprecated. The minimum required version of vCenter server will
be raised to 5.1 in the 8.0.0 release.

Change-Id: I38f520cace67ddcaf067787683ac47f50a6c0b3f