Updated openstack/openstack Project: openstack-infra/project-config  cd2bee83a164d8e77b096ca2441a7a58323f1fcc
Revert "zuul/layout: extend puppet-openstack-integration testing"

Quote from Paul Belanger:
"Sadly, this gate is not going to work as expected. They expect the local git repo to actually be puppet-keystone, since this is the puppet-openstack-integration module we don't ship any acceptance tests.
(...)"

It's pretty clear we need more work to get this done.
This reverts commit 5d6a86e7fbd862aabc7a0b67d1dff7de428520ba.

Change-Id: I843f2ecc1dda6a8353654e18fbf10c87bd162df8