Updated openstack/openstack Project: openstack/sahara  3801f5e7a5cedee60bde55f7434df65ac317afcb
Support placeholders in args of job for i/o

Several job types (at least spark) can require having url of input
or output data in arguments of job execution. The aim of the patch
is to add support of testing such types of job executions.
Now you can add {input_datasource} or {output_datasource}
into args of template of job execution, and that will be replaced
with corresponding urls.

Change-Id: I816ca6b559e3878c97f7d8d5088e0b458a19bd1b