Updated openstack/openstack Project: openstack/manila  baa83c54142968c93dbd1fa89e3fb5256b2d27ef
Support extend_share in NetApp cDOT drivers

Now that extend_share is in Manila core, we can support the new API in
the NetApp cDOT drivers.  As this patch adds methods to the driver shim
layer, also add a unit test to ensure the shims remain in sync for each
cDOT driver.

Implements: blueprint extend-share-in-netapp-cdot-drivers
Change-Id: I407a12a0cd44f37a62bd500251049c50a097f2c4