Updated openstack/openstack Project: openstack-dev/grenade  2eaf7fb2f82a8f93b143d8c3b091156251503d8b
Make ssh commands stateless

This makes our ssh-into-test-instances commands completely stateless. That
means they accept any key automatically (which we already did) and not
write them out to the known_hosts file of the user running grenade. Without
this, multiple runs of grenade (or existing host keys for the IPs we use)
will conflict and cause us to fail when validating things over ssh because
of the key mismatches.

Change-Id: I0be425764cd640ada4f00300167df77fd870214e