Updated openstack/openstack Project: openstack-infra/devstack-gate  37267b1314a425fc1292a7834ceecae7ec7f9086
Make it possible to not remove stack sudo

In cases where we are using devstack plugins to test services against
devstack that are not openstack services we may want to allow those
services to have sudo access. Prior to this change this was not possible
as d-g always removed the stack users sudo rights.

Add a flag that continues to disable sudo by default but allows for
stack user sudo access when appropriate.

Change-Id: I78ebf6432e9ea680cdb8bf103967cf9ed6c1b401