Updated openstack/openstack Project: openstack/manila  560fcf0d9f3e7f99153aebcea5bf49d3318a8c66
Fix HP3PAR extra-specs scoping prefix bug

The recently added HP 3PAR extra-specs for create share options
use the wrong prefix for scoping. The prefix should not have used
an underscore in it. In addition, the underscore used in the docs
was different than in the code.

The new scoping prefix will be "hpe3par" (to be followed by a colon
for scoping or underscore for future capabilities).

Fixes Bug: #1481106

Change-Id: Ibd326fc367c1845e56706969daf01359fa7b2dac