Updated openstack/openstack Project: openstack/nova  317e01c1a0c1616b18364c76d7854e647526b7b4
VMware: ensure that get_info raises the correct exception

In the event that a instance is deleted during the get_info method
then the method would not raise the correct exception.

If the instance no longer exists then the VC will raise
a 'ManagedObjectNotFound' exception.

Note that the oslo.vmware patch that supports this exception is:
https://review.openstack.org/#/c/191553/

This exception handling is treated by using vexc.register_fault_class
to support the local exception (until replaced by oslo.vmware)

Co-Authored-By: Radoslav Gerganov <rgerganov@vmware.com>

Change-Id: I2d4b131789319cfd1322cbc95edd7822114532e1
Closes-bug: #1464979