Updated openstack/openstack Project: openstack-infra/project-config  7f5400698e1cd112ef608c05b0e290b26c1af7a3
Remove rally-jobs job group

This job group is not really needed, in nearly all cases a single job of
the group is needed and never the whole group.

Instead add to each repository the single job that is used.

Exception: rally itself uses most of the job group, add all used jobs to
it.

This change is just a cleanup, it should not have any side effects on
which jobs run.
It reduces the number of created Jenkins jobs from 6521 to 6464 since
many unused jobs are not generated anymore.

Change-Id: I0a309aa975dccefbcc111c28a12aad9872681dcb